% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regchange.R
\name{regchange}
\alias{regchange}
\title{Inhomogeneous linear regression model}
\usage{
regchange(
  x,
  y,
  algorithm = "exact",
  kmax = 5,
  gamma = 0.01,
  lambda = seq(1, 0.01, -0.01),
  delta = 0,
  depth = 3,
  intercept = TRUE,
  train = seq(1, length(y)),
  subs = seq(1, length(y)),
  type = "glmnet"
)
}
\arguments{
\item{x}{matrix of covariates of dimension nxp}

\item{y}{response vector of length n}

\item{algorithm}{either \code{"exact"}  for the dynamic programming algorithm or \code{"binary"} for the approximate binary segmentation
algorithm}

\item{kmax}{maximum number of segments in the model}

\item{gamma}{penalized factor for the number of segments in the model}

\item{lambda}{vector of penalized parameters for the Lasso estimator on each segment}

\item{delta}{minimum proportion of sample points on each segment}

\item{depth}{maximum length of a leaf in the tree generated by the binary segmentation algorithm}

\item{intercept}{should the model have an intercept}

\item{train}{vector with train points}

\item{subs}{vector with candidate change-points}

\item{type}{either \code{"glmnet"} or \code{"lars"}}
}
\value{
a list containing the change point vector, a matrix with the regression parameters
on each segment and a matrix with the residual sum of squares for each value of \code{lambda} and number of considered
segments in the model
\item{alpha}{an array of dimension kmax x  kmax x length(lambda). For each value of lambda and number of segments j, the vector
alpha[, 1:j, lambda] contains the change-point proportions with respect to the sample size n. }
\item{beta}{a list of dimension kmax x length(lambda) containing in each entry the matrix of regression coefficients}
\item{rss}{matrix of dimension kmax x length(lambda) with the sum of squared residuals for each model}
}
\description{
This function fits a linear regression model with change-points in the regression coefficients by a regularized
least squares criterion. The algorithm computes the exact solution by using a dynamic programming approach.
}
\examples{
x <- matrix(runif(150), ncol=3, nrow=50)
y <- x\%*\%c(1,0,0)+rnorm(50, 0.01)
fit <- regchange(x,y)

}
\references{
Leonardi, F  and Bühlmann, P.  Computationally efficient change point detection for high-dimensional regression.
\href{https://arxiv.org/abs/1601.03704}{arXiv:1601.03704}
}
\author{
Florencia Leonardi and Peter Bühlmann \cr Maintainer: Florencia Leonardi \email{florencia@usp.br}
}
